update database and delete in controller

+++ src/main/java/anna/freimuth/repo/ItemRepo.java
@@ -3,6 +3,6 @@ package anna.freimuth.repo;
 import anna.freimuth.entity.Item;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ItemRepo extends CrudRepository<Item, Integer>{
+public interface ItemRepo extends CrudRepository<Item, Long>{
 
 }
